Update App.xaml.cs
- Refactored the whole thing so it specifically says XeniaCanary where it used to just say Xenia

+++ Xenia Manager/App.xaml.cs
@@ -2,6 +2,7 @@
 using System.Globalization;
 using System.IO;
 using System.Net.Http;
+using System.Reflection;
 using System.Runtime.InteropServices;
 using System.Windows;
 using Microsoft.Win32;
@@ -86,11 +87,11 @@ namespace Xenia_Manager
         /// If there is a newer version, ask user if he wants to update
         /// If user wants to update, update Xenia to the latest version
         /// </summary>
-        private async Task CheckForXeniaUpdates()
+        private async Task CheckForXeniaCanaryUpdates()
         {
             try
             {
-                Log.Information("Checking for updates.");
+                Log.Information("Checking for Xenia Canary updates");
 
                 using (HttpClient client = new HttpClient())
                 {
@@ -105,10 +106,10 @@ namespace Xenia_Manager
                         JObject latestRelease = JObject.Parse(json);
                         DateTime releaseDate;
                         DateTime.TryParseExact(latestRelease["published_at"].Value<string>(), "MM/dd/yyyy HH:mm:ss", CultureInfo.InvariantCulture, DateTimeStyles.None, out releaseDate);
-                        if (releaseDate != appConfiguration.Xenia.ReleaseDate)
+                        if (releaseDate != appConfiguration.XeniaCanary.ReleaseDate)
                         {
                             Log.Information("Found newer version of Xenia");
-                            MessageBoxResult result = MessageBox.Show("Found a new version of Xenia. Do you want to update Xenia?", "Confirmation", MessageBoxButton.YesNo);
+                            MessageBoxResult result = MessageBox.Show("Found a new version of Xenia Canary. Do you want to update it?", "Confirmation", MessageBoxButton.YesNo);
 
                             if (result == MessageBoxResult.Yes)
                             {
@@ -126,16 +127,16 @@ namespace Xenia_Manager
                                     downloadManager.downloadUrl = downloadUrl;
                                     downloadManager.downloadPath = AppDomain.CurrentDomain.BaseDirectory + @"\xenia.zip";
                                     Log.Information("Downloading the latest Xenia Canary build");
-                                    await downloadManager.DownloadAndExtractAsync();
+                                    await downloadManager.DownloadAndExtractAsync(appConfiguration.XeniaCanary.EmulatorLocation);
                                     Log.Information("Downloading and extraction of the latest Xenia Canary build done");
 
                                     // Update configuration
-                                    appConfiguration.Xenia.Version = (string)latestRelease["tag_name"];
-                                    appConfiguration.Xenia.ReleaseDate = releaseDate;
-                                    appConfiguration.Xenia.LastUpdateCheckDate = DateTime.Now;
+                                    appConfiguration.XeniaCanary.Version = (string)latestRelease["tag_name"];
+                                    appConfiguration.XeniaCanary.ReleaseDate = releaseDate;
+                                    appConfiguration.XeniaCanary.LastUpdateCheckDate = DateTime.Now;
                                     await appConfiguration.SaveAsync(AppDomain.CurrentDomain.BaseDirectory + "config.json");
-                                    Log.Information("Xenia has been updated to the latest build");
-                                    MessageBox.Show("Xenia has been updated to the latest build");
+                                    Log.Information("Xenia Canary has been updated to the latest build");
+                                    MessageBox.Show("Xenia Canary has been updated to the latest build");
                                 }
                             }
                         }
@@ -158,7 +159,7 @@ namespace Xenia_Manager
             finally
             {
                 // Always update last update check date
-                appConfiguration.Xenia.LastUpdateCheckDate = DateTime.Now;
+                appConfiguration.XeniaCanary.LastUpdateCheckDate = DateTime.Now;
                 await appConfiguration.SaveAsync(AppDomain.CurrentDomain.BaseDirectory + "config.json");
             }
         }
@@ -167,12 +168,15 @@ namespace Xenia_Manager
         {
             try
             {
-                Log.Information("Downloading Xenia Manager Updater");
-                await downloadManager.DownloadFileAsync("https://github.com/xenia-manager/xenia-manager/releases/download/updater/Xenia.Manager.Updater.zip", AppDomain.CurrentDomain.BaseDirectory + @"\xenia manager updater.zip");
-                Log.Information("Extracting Xenia Manager Updater");
-                downloadManager.ExtractZipFile(AppDomain.CurrentDomain.BaseDirectory + @"\xenia manager updater.zip", AppDomain.CurrentDomain.BaseDirectory);
-                Log.Information("Cleaning up");
-                downloadManager.DeleteFile(AppDomain.CurrentDomain.BaseDirectory + @"\xenia manager updater.zip");
+                if (!File.Exists(AppDomain.CurrentDomain.BaseDirectory + "Xenia Manager Updater.exe"))
+                {
+                    Log.Information("Downloading Xenia Manager Updater");
+                    await downloadManager.DownloadFileAsync("https://github.com/xenia-manager/xenia-manager/releases/download/updater/Xenia.Manager.Updater.zip", AppDomain.CurrentDomain.BaseDirectory + @"\xenia manager updater.zip");
+                    Log.Information("Extracting Xenia Manager Updater");
+                    downloadManager.ExtractZipFile(AppDomain.CurrentDomain.BaseDirectory + @"\xenia manager updater.zip", AppDomain.CurrentDomain.BaseDirectory);
+                    Log.Information("Cleaning up");
+                    downloadManager.DeleteFile(AppDomain.CurrentDomain.BaseDirectory + @"\xenia manager updater.zip");
+                }
                 await Task.Delay(1);
             }
             catch (Exception ex)
@@ -309,20 +313,39 @@ namespace Xenia_Manager
             // Checking if there is a configuration file
             if (appConfiguration != null)
             {
-                // If there is a configuration file, check if it already checked for Xenia updates
-                if (appConfiguration.Xenia.LastUpdateCheckDate == null || (DateTime.Now - appConfiguration.Xenia.LastUpdateCheckDate.Value).TotalDays >= 1)
-                {
-                    // If it didn't, check for a Xenia update
-                    await CheckForXeniaUpdates();
-                }
-
                 // Load the correct theme
                 await LoadTheme();
+
+                // Check if Xenia Canary is installed
+                if (appConfiguration.XeniaCanary != null)
+                {
+                    // Check if it already checked for Xenia Canary updates
+                    if (appConfiguration.XeniaCanary.LastUpdateCheckDate == null || (DateTime.Now - appConfiguration.XeniaCanary.LastUpdateCheckDate.Value).TotalDays >= 1)
+                    {
+                        // If it didn't, check for a Xenia update
+                        await CheckForXeniaCanaryUpdates();
+                    }
+                }
+                else
+                {
+                    // If there isn't configuration file, download Xenia Manager Updater and load Welcome Window
+                    WelcomeDialog welcome = new WelcomeDialog();
+                    welcome.Show();
+                }
             }
             else
             {
-                // If there isn't configuration file, download Xenia Manager Updater and load Welcome Window
+                // If there isn't Xenia installed, load Welcome Window
                 await DownloadXeniaManagerUpdater();
+                appConfiguration = new Configuration();
+                appConfiguration.ThemeSelected = "Light";
+                appConfiguration.Manager = new UpdateInfo
+                {
+                    Version = $"{Assembly.GetExecutingAssembly().GetName().Version.Major}.{Assembly.GetExecutingAssembly().GetName().Version.Minor}.{Assembly.GetExecutingAssembly().GetName().Version.Build}",
+                    ReleaseDate = null,
+                    LastUpdateCheckDate = DateTime.Now
+                };
+                await appConfiguration.SaveAsync(AppDomain.CurrentDomain.BaseDirectory + "config.json");
                 WelcomeDialog welcome = new WelcomeDialog();
                 welcome.Show();
             }
